+++ libexec/rtld-elf/mips/cheri_plt_code.S
