+++ libexec/rtld-elf/mips/rtld_small_vector.h
