+++ libexec/rtld-elf/riscv/rtld_machdep.h
@@ -2,6 +2,8 @@
  * Copyright (c) 1999, 2000 John D. Polstra.
  * Copyright (c) 2015 Ruslan Bukin <br@bsdpad.com>
  * All rights reserved.
+ * Copyright 2018-2020 Alex Richardson <arichardson@FreeBSD.org>
+ * Copyright 2020 Jessica Clarke <jrtc27@FreeBSD.org>
  *
  * Portions of this software were developed by SRI International and the
  * University of Cambridge Computer Laboratory under DARPA/AFRL contract
@@ -41,21 +43,121 @@
 #include <sys/types.h>
 #include <machine/atomic.h>
 
+#ifdef __CHERI_PURE_CAPABILITY__
+#include <cheri/cheri.h>
+#include <cheri/cheric.h>
+#endif
+
 struct Struct_Obj_Entry;
 
+#ifndef __CHERI_PURE_CAPABILITY__
 uint64_t set_gp(struct Struct_Obj_Entry *obj);
+#endif
 
 /* Return the address of the .dynamic section in the dynamic linker. */
+#ifdef __CHERI_PURE_CAPABILITY__
+#define rtld_dynamic(obj) (&_DYNAMIC)
+#else
 #define rtld_dynamic(obj)                                               \
 ({                                                                      \
 	Elf_Addr _dynamic_addr;                                         \
 	__asm __volatile("lla       %0, _DYNAMIC" : "=r"(_dynamic_addr));   \
 	(const Elf_Dyn *)_dynamic_addr;                                 \
 })
+#endif
 
 Elf_Addr reloc_jmpslot(Elf_Addr *where, Elf_Addr target,
     const struct Struct_Obj_Entry *defobj, const struct Struct_Obj_Entry *obj,
     const Elf_Rel *rel);
+
+#ifdef __CHERI_PURE_CAPABILITY__
+
+#define	FUNC_PTR_REMOVE_PERMS						\
+	(CHERI_PERM_SEAL | CHERI_PERM_STORE | CHERI_PERM_STORE_CAP |	\
+	CHERI_PERM_STORE_LOCAL_CAP)
+
+#define	DATA_PTR_REMOVE_PERMS						\
+	(CHERI_PERM_SEAL | CHERI_PERM_EXECUTE)
+
+/* TODO: we should have a separate member for .text/rodata */
+#define get_codesegment(obj) ((obj)->text_rodata_cap)
+
+/* TODO: ABIs with tight bounds */
+#define can_use_tight_pcc_bounds(obj) ((void)(obj), false)
+
+/*
+ * Create a pointer to a function.
+ * Important: this is not necessarily callable! For ABIs with tight bounds we
+ * need to load CGP first -> use make_function_pointer() instead.
+ */
+static inline dlfunc_t
+make_code_pointer(const Elf_Sym *def, const struct Struct_Obj_Entry *defobj,
+    bool tight_bounds, size_t addend)
+{
+	const void *ret = get_codesegment(defobj) + def->st_value;
+	/* Remove store and seal permissions */
+	ret = cheri_clearperm(ret, FUNC_PTR_REMOVE_PERMS);
+	if (tight_bounds) {
+		ret = cheri_setbounds(ret, def->st_size);
+	}
+	/*
+	 * Note: The addend is required for C++ exceptions since capabilities
+	 * for catch blocks point to the middle of a function.
+	 */
+	ret = cheri_incoffset(ret, addend);
+	/* All code pointers should be sentries: */
+	ret = __builtin_cheri_seal_entry(ret);
+	return __DECONST(dlfunc_t, ret);
+}
+
+/*
+ * Create a function pointer that can be called anywhere
+ */
+static inline dlfunc_t
+make_function_pointer_with_addend(
+    const Elf_Sym *def, const struct Struct_Obj_Entry *defobj, size_t addend)
+{
+	/* TODO: ABIs with tight bounds */
+	return make_code_pointer(def, defobj, /*tight_bounds=*/false, addend);
+}
+
+static inline dlfunc_t
+make_function_pointer(const Elf_Sym *def, const struct Struct_Obj_Entry *defobj)
+{
+	return make_function_pointer_with_addend(def, defobj, /*addend=*/0);
+}
+
+static inline void*
+make_data_pointer(const Elf_Sym *def, const struct Struct_Obj_Entry *defobj)
+{
+	void *ret = defobj->relocbase + def->st_value;
+	/* Remove execute and seal permissions */
+	ret = cheri_clearperm(ret, DATA_PTR_REMOVE_PERMS);
+	/* TODO: can we always set bounds here or does it break compat? */
+	ret = cheri_setbounds(ret, def->st_size);
+	return ret;
+}
+
+#define set_bounds_if_nonnull(ptr, size)	\
+	do { if (ptr) { ptr = cheri_setbounds_sametype(ptr, size); } } while(0)
+
+/* ignore _init/_fini */
+#define call_initfini_pointer(obj, target) rtld_fatal("%s: _init or _fini used!", obj->path)
+
+/* TODO: Per-function captable/PLT/FNDESC support (needs CGP) */
+#define call_init_array_pointer(obj, target)				\
+	(((InitArrFunc)(target).value)(main_argc, main_argv, environ))
+
+#define call_fini_array_pointer(obj, target)				\
+	(((InitFunc)(target).value)())
+
+/*
+ * TODO: Not implemented for CHERI.
+ * #define	call_ifunc_resolver(ptr) \
+ * 	(((Elf_Addr (*)(void))ptr)())
+ */
+
+#else /* __CHERI_PURE_CAPABILITY__ */
 
 #define make_function_pointer(def, defobj) \
 	((defobj)->relocbase + (def)->st_value)
@@ -79,12 +181,18 @@ Elf_Addr reloc_jmpslot(Elf_Addr *where, Elf_Addr target,
 #define	call_ifunc_resolver(ptr) \
 	(((Elf_Addr (*)(void))ptr)())
 
+#endif /* __CHERI_PURE_CAPABILITY__ */
+
 /*
  * TLS
  */
 #define	TLS_TP_OFFSET	0x0
+#ifdef __CHERI_PURE_CAPABILITY__
+#define	TLS_DTV_OFFSET	0
+#else
 #define	TLS_DTV_OFFSET	0x800
-#define	TLS_TCB_SIZE	16
+#endif
+#define	TLS_TCB_SIZE	(2 * sizeof(void *))
 
 #define round(size, align) \
     (((size) + (align) - 1) & ~((align) - 1))
@@ -112,13 +220,34 @@ extern void *__tls_get_addr(tls_index* ti);
 static inline bool
 _rtld_validate_target_eflags(const char *path, Elf_Ehdr *hdr, const char *main_path)
 {
-	if (hdr->e_flags & EF_RISCV_CHERIABI) {
-		_rtld_error("%s: cannot load %s since it is CheriABI",
-		    main_path, path);
+	bool rtld_is_cheriabi, hdr_is_cheriabi;
+
+#ifdef __CHERI_PURE_CAPABILITY__
+	rtld_is_cheriabi = true;
+#else
+	rtld_is_cheriabi = false;
+#endif
+	hdr_is_cheriabi = (hdr->e_flags & EF_RISCV_CHERIABI) != 0;
+
+	if (rtld_is_cheriabi != hdr_is_cheriabi) {
+		_rtld_error("%s: cannot load %s since it is%s CheriABI",
+		    main_path, path, hdr_is_cheriabi ? "" : " not");
 		return (false);
 	}
 
 	return (true);
 }
+
+#ifdef __CHERI_PURE_CAPABILITY__
+static inline void
+fix_obj_mapping_cap_permissions(Obj_Entry *obj, const char *path __unused)
+{
+	obj->text_rodata_cap = (const char*)cheri_clearperm(obj->text_rodata_cap, FUNC_PTR_REMOVE_PERMS);
+	obj->relocbase = (char*)cheri_clearperm(obj->relocbase, DATA_PTR_REMOVE_PERMS);
+	obj->mapbase = (char*)cheri_clearperm(obj->mapbase, DATA_PTR_REMOVE_PERMS);
+	/* Purecap code also needs the capmode flag */
+	obj->text_rodata_cap = cheri_setflags(obj->text_rodata_cap, CHERI_FLAGS_CAP_MODE);
+}
+#endif
 
 #endif
